- Update Save Cookie with all contract information after customer register. Link to payment View - Modify Customer.css file.

+++ VehicleInsuranceClient/Program.cs
@@ -5,7 +5,7 @@ namespace VehicleInsuranceClient
     public class Program
     {
         public static string ApiAddress = "https://localhost:7008/api";
-        public static short CookieEstimateDuration = 14;
+        public static double CookieEstimateDuration = 14;
         public static void Main(string[] args)
         {
             var builder = WebApplication.CreateBuilder(args);
